refactored types file
