+++ imdb_project/src/contexts/UsersContext.tsx
@@ -1,6 +1,6 @@
 import { createContext, useState, useReducer, useEffect } from "react";
 
-import { User, ChildrenElementType, UsersContextTypes, UsersReducerActionTypes } from "../types";
+import { User, ChildrenProp, UsersContextTypes, UsersReducerActionTypes } from "../types";
 
 const reducer = (state: User[], action: UsersReducerActionTypes): User[] => {
   switch (action.type) {
@@ -22,7 +22,7 @@ const reducer = (state: User[], action: UsersReducerActionTypes): User[] => {
 
 const UsersContext = createContext<UsersContextTypes | undefined>(undefined);
 
-const UsersProvider = ({ children }: ChildrenElementType) => {
+const UsersProvider = ({ children }: ChildrenProp) => {
 
   const [users, dispatch] = useReducer(reducer, []);
   const [loggedInUser, setLoggedInUser] = useState<User | null>(() => {
