+++ imdb_project/src/main.tsx
@@ -2,7 +2,7 @@ import { createRoot } from 'react-dom/client';
 import './index.css';
 import App from './App.tsx';
 import { BrowserRouter } from 'react-router';
-import { UsersProvider } from './context/UsersContext.tsx';
+import { UsersProvider } from './contexts/UsersContext.tsx';
 import { MoviesProvider } from './contexts/MoviesContext.tsx';
 
 createRoot(document.getElementById('root') as HTMLDivElement).render(
