+++ imdb_project/src/types.ts
@@ -1,3 +1,4 @@
+import { ActionTypes } from "./contexts/MoviesContext"
 export type User = {
   id: string;
   username: string;
@@ -7,7 +8,7 @@ export type User = {
   profilePicture?: string;
   role: 'customer' | 'admin';
 };
-export type ChildrenElementType = {
+export type ChildrenProp = {
   children: React.ReactElement
 }
 export type UsersReducerActionTypes = 
@@ -25,4 +26,74 @@ export interface LoginValues {
   email: string;
   password: string;
   stayLoggedIn: boolean;
+}
+type IMDBRating = {
+  score: number,
+  votes: number
+}
+
+type IMDBData = {
+  totalScore: number,
+  userRatings?: IMDBRating[]
+}
+
+type MoviePopularity = {
+  ranking: number,
+  weeklyChange: number
+}
+
+type MoviePhotos = {
+  poster: string[],
+  cutscenes: string[]
+}
+
+type MovieVideos = {
+  trailers: string[],
+  cutscenes: string[]
+}
+
+type MovieWriter = {
+  name: string,
+  role: string
+}
+
+type MovieActor = {
+  name: string,
+  character: string[],
+  actorPhoto: string
+}
+
+type CastAndCrew = {
+  director: string,
+  writers: MovieWriter[],
+  actors: MovieActor[]
+}
+
+type MovieReviews = {
+  users: number,
+  critics: number,
+  metascore: number
+}
+
+export type Movie = {
+  id: string,
+  title: string,
+  releaseYear: number,
+  eirinCategory: string,
+  length: number,
+  IMDB: IMDBData,
+  popularity?: MoviePopularity,
+  photos: MoviePhotos,
+  videos: MovieVideos,
+  genres: string[],
+  description: string,
+  castAndCrew: CastAndCrew,
+  reviews?: MovieReviews
+}
+
+export type MoviesContextTypes = {
+  movies: Movie[],
+  dispatch: React.ActionDispatch<[action: ActionTypes]>
+  addNewMovie: (newMovie: Movie) => void,
+  findProduct: (id: Movie["id"]) => Movie | string
 }
